bug_reporter.c: suppress warnings
* ext/-test-/bug_reporter/bug_reporter.c (sample_bug_reporter): cast
  via uintptr_t to suppress warning cast between pointer and different
  size integer by gcc 4.8.

* ext/-test-/bug_reporter/bug_reporter.c (register_sample_bug_reporter):
  ditto.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@43339 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/-test-/bug_reporter/bug_reporter.c
@@ -6,14 +6,14 @@ int rb_bug_reporter_add(void (*func)(FILE *, void *), void *data);
 static void
 sample_bug_reporter(FILE *out, void *ptr)
 {
-    int n = (int)ptr;
+    int n = (int)(uintptr_t)ptr;
     fprintf(out, "Sample bug reporter: %d\n", n);
 }
 
 static VALUE
 register_sample_bug_reporter(VALUE self, VALUE obj)
 {
-    rb_bug_reporter_add(sample_bug_reporter, (void *)NUM2INT(obj));
+    rb_bug_reporter_add(sample_bug_reporter, (void *)(uintptr_t)NUM2INT(obj));
     return Qnil;
 }
